requirements.txt
aiohttp==3.7.4.post0
botbuilder-integration-aiohttp>=4.14.0
botbuilder-dialogs>=4.14.0
botbuilder-ai>=4.14.0
datatypes-date-time>=1.0.0.a2
python-dotenv~=0.21.0
botbuilder-applicationinsights>=4.14.0
botbuilder-integration-applicationinsights-aiohttp>=4.14.0